backend/requirements.txt
hume[evi]
livekit-agents
livekit-plugins-hume
livekit-plugins-openai
livekit-plugins-groq
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-noise-cancellation
python-dotenv
requests
fastapi
uvicorn
livekit